ステージをビルで囲んだ

+++ Framework/Easing.cpp
@@ -17,7 +17,7 @@ namespace Easing
 	{
 		if (targetTime < time) return 1;
 		time /= targetTime;
-		return 1 - cos((time * PI) / 2);
+		return 1 - static_cast<float>(cos((time * PI) / 2));
 	}
 
 	float InQuad(float time, float targetTime)
@@ -46,7 +46,7 @@ namespace Easing
 	{
 		if (targetTime < time) return 1;
 		time /= targetTime;
-		return sin((time * PI) / 2);
+		return static_cast<float>(sin((time * PI) / 2));
 	}
 
 	float OutQuad(float time, float targetTime)
@@ -60,14 +60,14 @@ namespace Easing
 	{
 		if (targetTime < time) return 1;
 		time /= targetTime;
-		return 1 - pow(1 - time, 3);
+		return 1 - static_cast<float>(pow(1 - time, 3));
 	}
 
 	float OutQuart(float time, float targetTime)
 	{
 		if (targetTime < time) return 1;
 		time /= targetTime;
-		return 1 - pow(1 - time, 4);
+		return 1 - static_cast<float>(pow(1 - time, 4));
 	}
 
 
@@ -75,7 +75,7 @@ namespace Easing
 	{
 		if (targetTime < time && !isReturn) return 1;
 		time /= targetTime;
-		return -(cos(PI * time) - 1) / 2;
+		return - static_cast<float>((cos(PI * time) - 1) / 2);
 	}
 
 	float InOutQuad(float time, float targetTime, bool isReturn)
@@ -83,21 +83,21 @@ namespace Easing
 		if (targetTime < time && !isReturn) return 1;
 
 		time /= targetTime;
-		return time < 0.5 ? 2 * time * time : 1 - pow(-2 * time + 2, 2) / 2;
+		return time < 0.5 ? 2 * time * time : 1 - static_cast<float>(pow(-2 * time + 2, 2) / 2);
 	}
 
 	float InOutCubic(float time, float targetTime, bool isReturn)
 	{
 		if (targetTime < time && !isReturn) return 1;
 		time /= targetTime;
-		return time < 0.5 ? 4 * time * time * time : 1 - pow(-2 * time + 2, 3) / 2;
+		return time < 0.5 ? 4 * time * time * time : 1 - static_cast<float>(pow(-2 * time + 2, 3) / 2);
 	}
 
 	float InOutQuart(float time, float targetTime, bool isReturn)
 	{
 		if (targetTime < time && !isReturn) return 1;
 		time /= targetTime;
-		return time < 0.5 ? 8 * time * time * time * time : 1 - pow(-2 * time + 2, 4) / 2;
+		return time < 0.5 ? 8 * time * time * time * time : 1 - static_cast<float>(pow(-2 * time + 2, 4) / 2);
 	}
 
 	DirectX::SimpleMath::Vector3 InOutCubic(const DirectX::SimpleMath::Vector3& start, const DirectX::SimpleMath::Vector3& end, float t)
@@ -118,7 +118,7 @@ namespace Easing
 			t = 8 * t * t * t * t; // ‘O”¼
 		}
 		else {
-			t = 1 - std::pow(-2 * t + 2, 4) / 2; // Œã”¼
+			t = 1 - static_cast<float>(std::pow(-2 * t + 2, 4) / 2); // Œã”¼
 		}
 		return DirectX::SimpleMath::Vector3::Lerp(start, end, t);
 	}
