+++ Game/Particle/ExplosionEffect.cpp
@@ -52,7 +52,9 @@ void ExplosionEffect::Set(DirectX::SimpleMath::Vector3 pos)
 	for (int i = 0; i < MAX_EFFECT_NUM; i++)
 	{
 		// ƒ|ƒWƒVƒ‡ƒ“‚Æ¶‘¶ŽžŠÔ‚ðÝ’è
-		m_posLife.insert({ pos + Vector3(rand() % 3,rand() % 3,rand() % 3),0.0f - i * 0.1f});
+		m_posLife.insert({ pos + Vector3(static_cast<float>(rand() % 3),static_cast<float>(rand() % 3),static_cast<float>(rand() % 3))
+			,0.0f - i * 0.1f}
+		);
 	}
